pyproject.toml
[project]
name = "tasker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14.0"
dependencies = [
    "fastapi>=0.122.1",
    "loguru>=0.7.3",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
    "sqlalchemy>=2.0.44",
    "structlog>=25.5.0",
    "uvicorn>=0.38.0",
]
